client/requirements.txt
# Client dependencies - AgentK MCP Client Application
# Web interface framework
streamlit==1.45.0

# OpenAI client for LLM integration
openai==1.77.0

# Model Context Protocol client
mcp==1.7.1

# HTTP client for async operations
httpx==0.28.1
httpx-sse==0.4.0
httpcore==1.0.9

# Utilities
python-dotenv==1.1.0
pytz==2025.2